+++ Adjutant/Blam/Halo2/TagAddressTranslator.cs
@@ -18,14 +18,14 @@ namespace Adjutant.Blam.Halo2
             this.cache = cache;
         }
 
-        public int GetAddress(int pointer)
+        public long GetAddress(long pointer)
         {
-            return pointer - Magic;
+            return (int)pointer - Magic;
         }
 
-        public int GetPointer(int address)
+        public long GetPointer(long address)
         {
-            return address + Magic;
+            return (int)address + Magic;
         }
     }
 }
