+++ Adjutant/Blam/Halo4/TagAddressTranslator.cs
@@ -26,14 +26,14 @@ namespace Adjutant.Blam.Halo4
             this.cache = cache;
         }
 
-        public int GetAddress(int pointer)
+        public long GetAddress(long pointer)
         {
-            return pointer - Magic;
+            return (int)pointer - Magic;
         }
 
-        public int GetPointer(int address)
+        public long GetPointer(long address)
         {
-            return address + Magic;
+            return (int)address + Magic;
         }
     }
 }
